restore spca behavior
